refactor: extract canvas logic into hook
